+++ mockall_derive/src/mock_trait.rs
@@ -1,5 +1,5 @@
 // vim: tw=80
-use quote::{ToTokens, quote};
+use quote::{ToTokens, format_ident, quote};
 use syn::{
     *,
     spanned::Spanned
@@ -12,18 +12,17 @@ use crate::{
 
 pub(crate) struct MockTrait {
     pub attrs: Vec<Attribute>,
-    pub consts: Vec<TraitItemConst>,
+    pub consts: Vec<ImplItemConst>,
     pub struct_generics: Generics,
-    pub trait_generics: Generics,
     pub methods: Vec<MockFunction>,
-    pub name: Ident,
+    pub path: Path,
     structname: Ident,
-    pub types: Vec<TraitItemType>
+    pub types: Vec<ImplItemType>
 }
 
 impl MockTrait {
     pub fn name(&self) -> &Ident {
-        &self.name
+        &self.path.segments.last().unwrap().ident
     }
 
     /// Create a new MockTrait
@@ -31,52 +30,55 @@ impl MockTrait {
     /// # Arguments
     /// * `structname` - name of the struct that implements this trait
     /// * `struct_generics` - Generics of the parent structure
-    /// * `trait_`  -    Mockable ItemTrait
+    /// * `impl_`  -    Mockable ItemImpl for a trait
     /// * `vis`     -   Visibility of the struct
     pub fn new(structname: &Ident,
                struct_generics: &Generics,
-               trait_: ItemTrait,
+               impl_: ItemImpl,
                vis: &Visibility) -> Self
     {
         let mut consts = Vec::new();
         let mut methods = Vec::new();
         let mut types = Vec::new();
-        for ti in trait_.items.into_iter() {
-            match ti {
-                TraitItem::Const(tic) => {
-                    if tic.default.is_none() {
-                        compile_error(tic.span(),
-                            "When mocking trait associated constants, a default value must be given");
-                    }
-                    consts.push(tic);
+        let path = if let Some((_, path, _)) = impl_.trait_ {
+            path
+        } else {
+            compile_error(impl_.span(), "impl block must implement a trait");
+            Path::from(format_ident!("__mockall_invalid"))
+        };
+        let ident = &path.segments.last().unwrap().ident;
+
+        for ii in impl_.items.into_iter() {
+            match ii {
+                ImplItem::Const(iic) => {
+                    consts.push(iic);
                 },
-                TraitItem::Method(tim) => {
-                    let mf = mock_function::Builder::new(&tim.sig, &vis)
-                        .attrs(&tim.attrs)
+                ImplItem::Method(iim) => {
+                    let mf = mock_function::Builder::new(&iim.sig, &vis)
+                        .attrs(&iim.attrs)
                         .levels(2)
                         .call_levels(0)
                         .struct_(structname)
                         .struct_generics(struct_generics)
-                        .trait_(&trait_.ident)
+                        .trait_(ident)
                         .build();
                     methods.push(mf);
                 },
-                TraitItem::Type(tit) => {
-                    types.push(tit);
+                ImplItem::Type(iit) => {
+                    types.push(iit);
                 },
                 _ => {
-                    compile_error(ti.span(),
+                    compile_error(ii.span(),
                     "This impl item is not yet supported by MockAll");
                 }
             }
         }
         MockTrait {
-            attrs: trait_.attrs,
+            attrs: impl_.attrs,
             consts,
             struct_generics: struct_generics.clone(),
-            trait_generics: trait_.generics,
             methods,
-            name: trait_.ident,
+            path,
             structname: structname.clone(),
             types
         }
@@ -92,7 +94,6 @@ impl MockTrait {
     pub fn trait_impl(&self, modname: &Ident) -> impl ToTokens {
         let attrs = &self.attrs;
         let (ig, tg, wc) = self.struct_generics.split_for_impl();
-        let (_, t_tg, _) = self.trait_generics.split_for_impl();
         let consts = &self.consts;
         let calls = self.methods.iter()
                 .map(|meth| meth.call(Some(modname)))
@@ -105,12 +106,12 @@ impl MockTrait {
             .filter(|meth| !meth.is_static())
             .map(|meth| meth.expect(&modname))
             .collect::<Vec<_>>();
-        let name = &self.name;
+        let path = &self.path;
         let structname = &self.structname;
         let types = &self.types;
         quote!(
             #(#attrs)*
-            impl #ig #name #t_tg for #structname #tg #wc {
+            impl #ig #path for #structname #tg #wc {
                 #(#consts)*
                 #(#types)*
                 #(#calls)*